rename use-new-activity-mutation
